Add files via upload
This repository presents a machine learning project that utilizes the Random Forest algorithm to predict brain tumors with high accuracy. After experimenting with three different algorithms, including Support Vector Machines (SVM) and Neural Networks (NN), we found that the Random Forest algorithm outperforms the others in terms of accuracy and efficiency.

Dataset

The dataset used in this project consists of Magnetic Resonance Imaging (MRI) brain images, comprising tumor images and  non-tumor images. The dataset is balanced, which ensures that the model is not biased towards any particular class.

Methodology

The project involves the following steps:

Data preprocessing and augmentation
Feature extraction
Model training and evaluation
Hyperparameter tuning
We applied various preprocessing and augmentation techniques to the MRI brain images to increase the size of the dataset and prevent overfitting. The Random Forest algorithm was then trained on the preprocessed data, and its performance was evaluated using metrics such as accuracy, precision, recall, and F1-score.

Results

The Random Forest algorithm achieved an accuracy of 96%, outperforming the SVM and NN algorithms. The model's high performance can be attributed to its ability to capture complex patterns and features in the input images.

Conclusion

This project demonstrates the effectiveness of the Random Forest algorithm in predicting brain tumors from MRI images. The model's high accuracy and efficiency make it a promising tool for medical diagnosis and research.