Create README.md
This project aims to develop a machine learning model that can accurately predict the presence of a brain tumor based on MRI images. We compare the performance of three algorithms: Random Forest, Support Vector Machine (SVM), and Neural Networks (NN), to identify the best approach for this task.

Objective was:

To develop a robust machine learning model for brain tumor disease prediction
To compare the performance of Random Forest, SVM, and NN algorithms on this task
To identify the best algorithm for brain tumor disease prediction
Dataset:

The dataset consists of MRI images of brains, divided into two classes: "no" (no tumor) and "yes" (tumor present).

Methodology:

Data Preprocessing: Load images from folders, resize to 100x100 pixels, and convert to grayscale.
Split Dataset: Split the dataset into training (80%) and testing sets (20%).
Random Forest: Train a Random Forest classifier on the training set and evaluate its performance on the testing set.
SVM: Train an SVM classifier on the training set and evaluate its performance on the testing set.
Neural Networks: Train a Neural Network classifier on the training set and evaluate its performance on the testing set.
Comparison: Compare the performance of the three algorithms using accuracy, precision, recall, F1 score, and confusion matrix.
Visualization: Visualize the results using confusion matrix and bar plots for evaluation metrics.
Results:
The accuracies of the
Random Forest model is 0.94, SVM model is 0.84,Neural Network model is 0.84.
The comparison of the three algorithms shows that 0.94 performs best on this task.

Conclusion:
This project demonstrates a comprehensive approach to brain tumor disease prediction using machine learning. By comparing the performance of Random Forest, SVM, and Neural Networks, we identify random forest the best algorithm for this task. The results of this project can be used to develop a robust and accurate brain tumor disease prediction system.